Youtube-DL Version
As of today (1/15/2023), I found a new library for downloading media and have decided to update the code to use yt-dlp instead. It has more features and fixes that youtube-dl does not seem to push out often.

This change will result in two different libraries being used, and I would like to keep the Youtube-DL version for archival purposes.