Major Folder Structure Changes
I do not want to get rid of the version that uses Youtube-DL, so I decided to create folders for each of the scripts that exist.

I will obvi recommend people use the yt-dlp version as it is the most up to date when it comes to websites and features.