Update to versions & cookies
This update removes the (in my opinion stupid) way I was handling creating separate versions of the script.

I remembered why there are 2 versions to begin with:

The way Windows and Linux process LF and CRLF.

The Linux release allows you to "Run in Console" without having to open a Console and type "python3" before the filename.

This update also includes a text file to pass cookies to the script. More details to come soon.